requirements.txt
Flask==3.0.3
mysql-connector-python==9.1.0
face-recognition==1.3.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
python-dotenv==1.0.1
Werkzeug==3.0.4
